Restructured rav
