+++ assets/js/base/rav/elems.js
@@ -2,5 +2,5 @@ import 'bootstrap/dist/js/bootstrap.bundle.min.js';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import {
   databaseAnnotatedGames
-} from './rav_elements.js';
-import '../../styles/app.css';
+} from './init.js';
+import '../../../styles/app.css';
