+++ assets/js/database/annotated_games/index.js
@@ -1,7 +1,7 @@
 import {
   databaseAnnotatedGames,
   ravMovesTable
-} from '../../../js/base/rav_elements.js';
+} from '../../../js/base/rav/init.js';
 
 databaseAnnotatedGames.modal.show();
