+++ templates/base/rav.html.twig
@@ -14,7 +14,7 @@
     {% block stylesheets %}{% endblock %}
     {% block javascripts %}
       {% block importmap %}
-        {{ importmap('js/base/rav.js') }}
+        {{ importmap('js/base/rav/elems.js') }}
       {% endblock %}
     {% endblock %}
   </head>
